Hydrate mission control traces from NullWatch
