+++ src/root.zig
@@ -17,8 +17,9 @@ pub const manager = @import("supervisor/manager.zig");
 pub const managed_skills = @import("managed_skills.zig");
 pub const meta_api = @import("api/meta.zig");
 pub const mdns = @import("mdns.zig");
+pub const mission_control = @import("core/mission_control.zig");
 pub const mission_control_api = @import("api/mission_control.zig");
-pub const mission_control_replay = @import("api/mission_control_replay.zig");
+pub const mission_control_replay = @import("core/mission_control_replay.zig");
 pub const observability_api = @import("api/observability.zig");
 pub const orchestrator = @import("installer/orchestrator.zig");
 pub const manifest = @import("core/manifest.zig");
@@ -67,6 +68,7 @@ test {
     _ = managed_skills;
     _ = meta_api;
     _ = mdns;
+    _ = mission_control;
     _ = mission_control_api;
     _ = mission_control_replay;
     _ = observability_api;
