+++ ui/src/lib/api/client.ts
@@ -1,4 +1,5 @@
 import { createOrchestrationApi } from '$lib/api/orchestration';
+import { createMissionControlApi } from '$lib/api/missionControl';
 import { createNullTicketsApi } from '$lib/api/nulltickets';
 import { encodePathSegment } from '$lib/orchestration/routes';
 
@@ -15,146 +16,29 @@ function withQuery(path: string, params: Record<string, string | number | boolea
 }
 
 export { encodePathSegment };
+export type {
+  MissionControlAgent,
+  MissionControlComponentMapping,
+  MissionControlControls,
+  MissionControlEvent,
+  MissionControlFailure,
+  MissionControlGraphEdge,
+  MissionControlGraphNode,
+  MissionControlObservabilityMapping,
+  MissionControlPhase,
+  MissionControlRecovery,
+  MissionControlReplayArtifact,
+  MissionControlState,
+  MissionControlStatus,
+  MissionControlTelemetry,
+  MissionControlTraceRef,
+  MissionControlWorkflowMapping,
+} from '$lib/api/missionControl';
 
 export type LogSource = 'instance' | 'nullhub';
 export type ReportOption = { value: string; label: string };
 export type ReportTypeOption = ReportOption & { labels: string[] };
 export type ReportRepoOption = ReportOption & { repo: string };
-export type MissionControlStatus = 'idle' | 'running' | 'intervention_required' | 'completed';
-export type MissionControlPhase =
-  | 'idle'
-  | 'launching'
-  | 'research'
-  | 'coding'
-  | 'checkpoint'
-  | 'testing'
-  | 'failed'
-  | 'forking'
-  | 'patching'
-  | 'retesting'
-  | 'review'
-  | 'completed';
-export type MissionControlControls = {
-  can_launch: boolean;
-  can_recover: boolean;
-  can_reset: boolean;
-};
-export type MissionControlAgent = {
-  id: string;
-  role: string;
-  status: string;
-  current_step: string;
-};
-export type MissionControlGraphNode = {
-  id: string;
-  label: string;
-  kind: string;
-  status: string;
-};
-export type MissionControlGraphEdge = {
-  from: string;
-  to: string;
-  status: string;
-};
-export type MissionControlTraceRef = {
-  kind: 'span' | 'eval';
-  run_id: string | null;
-  trace_id: string | null;
-  span_id: string | null;
-  eval_key: string | null;
-  operation: string;
-};
-export type MissionControlEvent = {
-  at_ms: number;
-  source: string;
-  level: string;
-  title: string;
-  detail: string;
-  status: string;
-  trace: MissionControlTraceRef | null;
-};
-export type MissionControlTelemetry = {
-  runs: number;
-  spans: number;
-  evals: number;
-  errors: number;
-  total_tokens: number;
-  total_cost_usd: number;
-  verdict: string;
-};
-export type MissionControlFailure = {
-  run_id: string;
-  checkpoint_id: string;
-  failed_step: string;
-  error_message: string;
-  suggested_intervention: string;
-};
-export type MissionControlRecovery = {
-  run_id: string;
-  forked_from: string;
-  human_instruction: string;
-  status: string;
-};
-export type MissionControlState = {
-  schema_version: number;
-  mode: string;
-  scenario_id: string;
-  scenario_version: string;
-  generated_at_ms: number;
-  mission_id: string;
-  title: string;
-  status: MissionControlStatus;
-  phase: MissionControlPhase;
-  headline: string;
-  elapsed_ms: number;
-  progress: number;
-  active_run_id: string | null;
-  failed_run_id: string | null;
-  recovered_run_id: string | null;
-  controls: MissionControlControls;
-  agents: MissionControlAgent[];
-  graph: {
-    nodes: MissionControlGraphNode[];
-    edges: MissionControlGraphEdge[];
-  };
-  events: MissionControlEvent[];
-  telemetry: MissionControlTelemetry;
-  failure: MissionControlFailure | null;
-  recovery: MissionControlRecovery | null;
-};
-export type MissionControlComponentMapping = {
-  component: string;
-  role: string;
-  evidence: string[];
-};
-export type MissionControlWorkflowMapping = MissionControlComponentMapping & {
-  checkpoint_id: string;
-  failed_run_id: string;
-  recovered_run_id: string;
-  human_instruction: string;
-};
-export type MissionControlObservabilityMapping = MissionControlComponentMapping & {
-  failed_run_id: string;
-  recovered_run_id: string;
-  trace_ref_source: string;
-};
-export type MissionControlReplayArtifact = {
-  artifact_schema_version: number;
-  artifact_kind: string;
-  generated_at_ms: number;
-  replay_fixture_path: string;
-  scenario_id: string;
-  scenario_version: string;
-  mode: string;
-  snapshot: MissionControlState;
-  replay_fixture: unknown;
-  ecosystem_mapping: {
-    nulltickets: MissionControlComponentMapping;
-    nullboiler: MissionControlWorkflowMapping;
-    nullclaw: MissionControlComponentMapping;
-    nullwatch: MissionControlObservabilityMapping;
-  };
-};
 type InstanceStartOptions = {
   launch_mode?: string;
   verbose?: boolean;
@@ -373,11 +257,7 @@ export const api = {
       }),
     ),
 
-  getMissionControlState: () => request<MissionControlState>('/mission-control/state'),
-  getMissionControlReplay: () => request<MissionControlReplayArtifact>('/mission-control/replay'),
-  launchMissionControl: () => request<MissionControlState>('/mission-control/launch', { method: 'POST' }),
-  resetMissionControl: () => request<MissionControlState>('/mission-control/reset', { method: 'POST' }),
-  recoverMissionControl: () => request<MissionControlState>('/mission-control/recover', { method: 'POST' }),
+  ...createMissionControlApi(request),
 
   applyUpdate: (c: string, n: string) =>
     request<any>(`/instances/${c}/${n}/update`, { method: 'POST' }),
